Support hmac_secret rotation in jwt_refresh feature
This supports the use of the previous secret, but does not take
any actions to update it to the new secret (there isn't actions
you can take, as the token is submitted as a parameter, it's not
part of the session or a cookie).  However, as jwt refresh tokens
are only valid for 30 minutes by default, rotation to a new secret
will occur quickly.  The only exception is when
allow_refresh_with_expired_jwt_access_token? is set, in which case
the user would need to decide how long they want to keep the old
secret around to support old expired access tokens.