Support hmac_secret rotation in webauthn feature
This only checks that the submitted challenges match either the
current or previous hmac.  The only time you would need to support
long hmac rotation period for webauthn is if you wanted to allow
users to go to the webauthn setup/auth page and then actually
submit the form on that page much later.