Support hmac_secret rotation in the otp feature
This will allow OTP authentication if the OTP key was created
with hmac_old_secret.  However, since it cannot update the OTP
secret on the device, it calls the otp_valid_code_for_old_secret
configuration method, and the user can then record whatever
information is needed, and use it to inform the user that they
need to rotate their OTP.

This explicitly does not handle creating a new OTP key if the
hmac_secret has changed between when the setup form was displayed
and when it was submitted.  The OTP would then need to be rotated,
and it's best to avoid that.  The user will get an invalid key
error and can submit again to use the new hmac_secret.