Require that *_response methods actually return a response or redirect
Now that these are overridable, if they were overridden improperly,
that can result in unexpected behavior.  Depending on the route,
depending on what the method returned, it could be taken as a Roda
route response (such as a String setting the request body).  In other
cases, the return value would be ignored, and control might pass to
error handling code.

Raise a RuntimeError if a *_response configuration method is used
improperly, indicating that it is a bug in the configuration and
not in Rodauth itself.

This is implemented by having Feature#response define two methods,
both of which are now private, and using auth_private_method for
the overridable method.