Return error response instead of 404 when token is missing
When a page requiring a token is requested without a token, Rodauth
currently returns a 404 response. If Rodauth is used as a middleware
with `:next_if_not_found` plugin option enabled, this request would be
forwarded to the main app. This can be misleading for the developer,
because it might communicate that it's not a Rodauth route.

We instead return an error response that the required token is missing,
which makes things clearer to the developer, and also ensures the
request doesn't get forwarded to the main app.