Handle internal_request_configuration blocks in superclasses
internal_request_configuration blocks in superclasses are now
respected when setting up the internal request class for a
subclass.  Note that by design, the internal request class is
a subclass of the related Rodauth::Auth class, and not a
subclass of the related Rodauth::Auth's superclass's internal
request class. All internal_request_configuration blocks,
whether specified in a superclass or the current class, are
processed as if they were specified directly in the current
class.

Update CHANGELOG for some recent changes while here.