Add otp_unlock feature for unlocking TOTP authentication after it has been locked out
Previously, the otp feature locked out TOTP authentication after
5 failed attempts.  Lockouts are required, as by default TOTP
authentication is a 1 in 10^6 chance, so without a lockout,
is trivial to brute force TOTP authentication.  However, the otp
feature did not support a way for the user to unlock TOTP
authentication, it could only be unlocked by the administrator.

This supports TOTP unlocking by requiring multiple consecutive
valid TOTP authentications in a short time period.  By default,
you can unlock TOTP authentication in about 5 minutes, since
by default 3 successful authentications are required, and there
is a minimum time of 2 minutes between authentications.

This should be infeasible to brute force, since brute force
success is only a 1 in 10^18 chance, and an invalid authentication
during unlock forces a 15 minute delay, so there are only a
total of 96 failed unlock attempts possible per day.

This stores the OTP unlock information in a separate table,
making it easy to integrate into an existing Rodauth configuration
that uses the otp feature.