Add `account_from_id` for loading account by id
This is useful in when delivering emails in a background job, where the
background job is given the account ID. In that case, the Rodauth
instance needs to be initialized outside of a request, and account and
token loaded into it.

  rodauth = RodauthApp.rodauth.allocate
  rodauth.account_from_id(id)
  rodauth.account #=> { ... }

There are also a few places within Rodauth that can make use of this
method.