Allow sending webauthn modify emails outside of a request
In a background job, we need to allocate a Rodauth instance without any
Rack request information if we wanted to use it to render emails. There
the session will not be available, but the account can be loaded into the
Rodauth instance.

  rodauth = RodauthApp.rodauth.allocate
  rodauth.account_from_id(account_id)
  rodauth.webauthn_authenticator_added_email_body

In order for the webauthn modify emails to render successfully in this
case, `#webauthn_account_id` needs to take the loaded account if
available.