Allow disabling WebAuthn autofill UI
This is useful when the developer wants the ability to provide WebAuthn
authentication without asking for login, but considers the autofill UI
too magical for the end user, and would prefer to display an explicit
"login via passkey" button instead.

The autofill UI has become even more magical in recent browser/OS
versions, where the user is asked to authenticate via WebAuthn as soon
as the login page is loaded. Previously, Chrome would display a passkey
option in a dropdown after focusing the login field, but this is no
longer the case.