Make clear_session not call the scope's clear session if using JWTs for session in the jwt feature
Previously, if using Roda's session plugin and the Rodauth jwt
feature, when the Rodauth clear_session was called for a request
that should use JWT, it would call clear_session on the Roda
scope.  That is a mistake, because all Rodauth session
information should be stored in the JWT in that case, not in a
session cookie.

This changes things so that Road scope clear_session is not called
if JWTs are in use.  It also makes it so that clear session does
not call session.clear twice in the JWT case.