Make rodauth and r.rodauth call default_rodauth_name for the default configuration to use
This makes it easier to use multiple Rodauth configurations in
an application.  Previously, you had to specify the alternate
configuration name in every case where you were calling the method.
This allows you to set the configuration to use in a single place,
so you can call rodauth and r.rodauth without arguments, and it
will pick up the desired configuration.