Make rodauth.has_password? method public
I needed this method when integrating rodauth-omniauth into our project.

If an account is created via omniauth, it doesn't have a password. In
this case, I prefer to display "Create Password" instead of "Change
Password".

Additionally, if the omniauth provider is the last authentication
method, including the password, I aim to block to disconnect it.

Therefore, I believe this method is beneficial to have in public.