Specify fixed locals for rendered templates by default, disable with use_template_fixed_locals? false
This is designed to work with the fixed locals support added to
Roda, in applications that specify default fixed locals for
templates using the :template_opts option to Roda's render
plugin.  A recommended practice for Roda going forward is to
specify that templates do not accept local variables by default,
and force templates that accept local variables to specify which
local variables are accepted:

  plugin :render, template_opts: {
      default_fixed_locals: '()',
      extract_fixed_locals: true
    }

This change allows Rodauth to work with such a recommendation,
by specifying that Rodauth templates take the rodauth local
variable (and the button template takes the value and opts
local variables). In cases where the user is passing additional
local variables via the :locals option, the assumption is they
are using their own templates, in which case Rodauth shouldn't
specify the fixed locals, so Rodauth doesn't set fixed locals
in that case.  There may be other cases where Rodauth shouldn't
set fixed locals, and for those cases, the
use_template_fixed_locals? false configuration method can be used.

If the Roda and Tilt versions support fixed locals, this run the
tests with fixed locals on by default.