Make verify-account-resend page work if verify_account_resend_explanatory_text calls verify_account_email_recently_sent?
This can happen in Rodauth configurations that want to suppress
the notice if the account has not sent the email recently.

This makes verify_account_email_recently_sent? return falsely if
there is no associated account, such as when the page is visited
directly, instead of being displayed in response to an account
creation or login attempt.  Note that if there is no account,
the warning should be displayed, because at that point you do not
know whether the email was sent recently or not.