Avoid exceeding 4K cookie size limit by setting an upper limit on path size when using login_return_to_requested_location?
Returning to requested location is a nice to have, but not a hard
requirement, and few legitimate cases will exceed the limit. Not
having a limit allows an easy way to unauthenticated users to force
cookie too large exceptions in applications.

This adds a login_return_to_requested_location_max_path_size
configuration method to change the limit.