Allow usage with Roda's plain_hash_response_headers plugin and Rack 3+
This changes Rodauth to use lower-case headers internally.  For
backwards compatibility on Rack < 3, the lower-case headers are
switched to mixed-case headers.  This isn't needed on Rack 3+, because
on Rack 3+, response headers are always lower case.  Roda has been
handling this transparently, as it downcases lower case headers by
default on Rack 3+.  However, Roda offers a
plain_hash_response_headers plugin for better performance by using a
plain hash for response headers.  This makes Rodauth compatible with
that.

This adds set_response_header and convert_response_header_key
private methods for handling the key conversion inside Rodauth.

In the specs, since it is easier to go from mixed case to lower case
instead of lower case to mixed case, keep the specs in mixed case, and
add a header spec method to convert mixed case to lower case.  Also
add a CONTENT_TYPE_KEY constant, which is used for setting the header.