Remove use of rdoc/task in Rakefile
With rdoc moving out of stdlib in Ruby 3.5, this could cause issues
with usage of bundle exec in CI.  It's also an unnecessary require
in the vast majority of cases, slowing down the Rakefile.

Switch to using a normal task.  In addition to being faster, I think
this is easier to understand.  This just generates the documentation
into www/public/rdoc.  Doesn't seem necessary to have a separate
task to generate into rdoc.

Note that the task itself requires rdoc, so actually generating the
documentation make not work if you are using bundle exec.