Clear account tokens when an account change is made
Previously, the only account change that cleared all account tokens
was closing the account. If you reset your password, only the reset
password token was cleared, if you verified a login change, only the
verify login change token was cleared, etc. This changes the default
behavior to clear most account tokens when an account change is made.

The following account changes trigger clearing of tokens:

* change login
* close account
* reset password
* unlock account
* verify account

The following account tokens are cleared upon such actions:

* active sessions (other than logged in session)
* email auth
* jwt refresh (if not logged in)
* lockout (updates token if it exists)
* remember (creates and uses new remember token if logged in via remember token)
* reset password
* single session (if not logged in)
* verify account
* verify login change

This provides more secure behavior in some cases. Let's say you
notice something funny with your email. You request a password
change. However, then you realize someone has access to your email,
so you change your login. Previously, the password reset link for
the account is still valid after the email change, so an attacker
can still change the password on the account post login change.
With this commit, once the login has been changed, the reset
password token is no longer valid. Similarly, once the password has
been reset, the verify login change token is no longer valid.

While I think this makes for a more secure and more appropriate
default, it's possible that this behavior is not desirable for all
Rodauth installations. To allow for customization, a clear_tokens
configuration method is available. This takes a block that is
passed a symbol for the change being made. The user can override
the behavior for specific symbols. They can also call super with
the reason to get the default behavior of clearing all related
tokens.

This unfortunately requires a number of special cases so that it
does not break expected behavior. For active_sessions and
single_session, you don't want to clear a token used for the
current session, because otherwise changing your login would result
in a logout.  For lockout, you cannot clear the token, because that
would result in an account unlock, so instead the token is updated.
Remember is a mix between active_sessions/single_session and lockout,
where you want to clear the token if the session is not logged in
via remember, if the session is logged in via remember, you want
to update the session value, so that the current session stays
remembered.

For jwt_refresh, this does not clear refresh tokens if the session
is logged in, as we don't know which refresh token is for the
current session. Potentially that ability could be added in the
future, but this issue is simple to address by using the
active_sessions feature, so recommend that approach.